Added new class - Actions. Map changed to class-collector. Actions - class-method.

+++ src/com/brightside/entities/inanimates/Grass.java
@@ -6,7 +6,7 @@ import java.awt.*;
 public class Grass extends Entity {
 
     @Override
-    public Image getImage(short cellSize) {
+    public Image getImage(int cellSize) {
         return t.getImage("src\\com\\brightside\\images\\grass_" + cellSize + ".png");
     }
 }
